pyproject.toml
[project]
name = "hu-datasets"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    { name = "MJGGerdes" },
    {name = "mjg.gerdes", email = "mjg.gerdes@avans.nl"},
]
readme = "README.md"
dependencies =[
    "black==25.1.0",
    "isort==6.0.1",
    "loguru==0.7.2",
    "mypy==1.13.0",
    "pandas==2.2.2",
    "requests==2.32.3",
    "ruff==0.11.2",
    "tqdm==4.66.1",
    "pillow==10.4.0",
    "gdown==5.2.0",
    "wget==3.2"
]
requires-python = "==3.12.*"
license = {text = "MIT"}

[tool.setuptools.packages.find]
where = ["."]
include = ["hu_datasets*"]

[tool.setuptools.package-data]
hu_datasets = ["**/*.json"]

[tool.pdm]
distribution = false
